Pin dependencies in package.json
A new version of chart.js (2.6) has been released. Because we
are depending on ^2.3, the newest version that matches 2.x.x
will be pulled down on npm install.

The new version of chart.js caused some issues in the RSSI
chart, so pinning this dependency on 2.5.0 for now to avoid
problems. Also doing the same for our other dependencies.