Fix: Start/pause button sometimes in wrong state
Steps to reproduce:
1. Open a device by selecting it in the device selector
2. Press “Pause” to pause reading the RSSI levels
3. Open another device (or, actually the same device again) by selecting
   it in the device selector

The start/pause button still shows “Start” as if the reading is paused
but the reading is actually started again (as I had expected right after
selecting a device).

Press on the “Start” button in this state does change the button to show
“Pause” but the reading of the RSSI levels continues. After pressing
“Pause” now, reading the RSSI levels is paused. So after the steps to
reproduce above are executed users have to press the start/pause button
twice to pause.

This fix solves this by making the start/pause button always show
“Pause” after selecting a device and acting appropriatly (that is,
pressing it does pause the reading).